% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delta.to.atpct}
\alias{delta.to.atpct}
\title{Convert isotope delta notation to atom percent}
\usage{
delta.to.atpct(delta, R = NULL, isotope = "N15", fraction = FALSE)
}
\arguments{
\item{delta}{A vector of delta-values, in permilles}

\item{R}{The R value of the reference material on which the delta values are based (defaults to NULL)}

\item{isotope}{Instead of passing the R value as argument, one can also indicated the
isotope using a character string. This works for C-13 (PD-belemnite), N-15 (Air), O-18 (VSMOW) and H-2 (VSMOW).
The specific formatting does not matter, e.g. `C13', `C-13', `13C' or `13-C' will all work.
`D' is equivalent to `H-2'.}

\item{fraction}{logical (defaults to FALSE).
If TRUE, then the isotopic content is returned as a fraction, i.e. 1.0 for 100\%.
If FALSE, a value of 100 indicates 100\%.}
}
\value{
atom percentages (or fractions)
}
\description{
Often, one needs to convert delta-permil data from isotope-ratio
mass spectrometry to atom-percent, i.e. the percentage (or fraction)
an isotopes makes up of a given element. This is crucial for mass balance
calculations, since delta-values are not proportional to the extent of
labelling.
}
\examples{
delta.to.atpct(-28,isotope="C13")
## [1] 1.080435
delta.to.atpct(c(23,47),isotope="N-15",fraction=TRUE)
## [1] 0.003746932 0.003834500
delta.to.atpct(-280,isotope="D")
## [1] 0.01121346
delta.to.atpct(-280,isotope="H-2")
## [1] 0.01121346
delta.to.atpct(-280,isotope="2H")
## [1] 0.01121346
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}

