% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethdf.R
\name{get_hdf_list}
\alias{get_hdf_list}
\alias{get_hdf_file}
\title{Retrieve MODIS HDF file}
\usage{
get_hdf_list(tiles = c("h18v04"), sat = c("aqua", "terra"),
  dates = c("0000.00.00", "9999.99.99"), version = 6,
  product = "13Q1", server = "http://e4ftl01.cr.usgs.gov",
  verbose = TRUE)

get_hdf_file(url = NULL, file = NULL, username = "user",
  password = "passw", overwrite = FALSE)
}
\arguments{
\item{tiles}{the MODIS tiles to download in format \code{h??v??}
(e.g. \code{h18v04})}

\item{sat}{the MODIS instruments for which data is retrieved,
e.g. \code{aqua} or \code{terra}, or both (default).}

\item{dates}{a date interval to which hdf file retrieval should be
restricted. Start and end dates are provided as
\code{yyyy.mm.dd}.  To retrieve the entire data of year 2010,
one could use \code{dates=c('2010.00.00','2010.99.99')}.}

\item{version}{MODIS file version, currently 6 (default).}

\item{product}{product to retrieve, defaults to \code{13Q1}.}

\item{server}{server to use, defaults to \code{http://e4ftl01.cr.usgs.gov}.}

\item{verbose}{logical indication whether progress should be logged
to the console (defaults to \code{TRUE}).}

\item{url}{full url of MODIS file to be downloaded.}

\item{file}{if specifies, \code{file} allows to change the file
name under which the MODIS hdf file is saved.}

\item{username, password}{username and password for hdf file
retrieval.}

\item{overwrite}{\code{get_hdf_file} skips the download of files
that already exist on disk, unless \code{overwrite=TRUE}.}
}
\value{
A data frame with columns \code{url}, \code{dir}, and
    \code{file}.
}
\description{
These functions retrieve MODIS Aqua and Terra hdf files and save
these locally.
}
\details{
\code{get_hdf_list} retrieves the URL of all MODIS hdf files from
Aqua, Terra, or both and returns them as a \code{data.frame}.  It
is possibly to specify a MODIS file version (currently 6, the
default) and to narrow down the list by specifying a date interval
(argument \code{dates}).

\code{get_hdf_file} retrieves on or more HDF files.
\code{username} and \code{password} to access the server need to be
passed as arguments.  Data retrieval occurs using \code{curl},
which is invoked using \code{system}.  Note that username and
password are temporarily stored in a local 'netrc' file, which is
removed after the hdf file has been downloaded.  This 'netrc' file
is readable by the user only (\code{chmod 0600}) and removed after
the hdf file has been downloaded.
}
\examples{
\dontrun{
require(pgeo);
h <- get_hdf_list(dates=c("2010.00.00","2010.01.14"))
get_hdf_files(h$url, username='user', password='secret')
}

}
