% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileMapper.R
\name{toMODISxy}
\alias{toMODISxy}
\alias{fromMODISxy}
\title{Convert MODIS (line,sample) to (latitude,longitude) and vice versa}
\usage{
toMODISxy(lat, lon = NULL, grid = 250)

fromMODISxy(gx, gy = NULL, grid = 250)
}
\arguments{
\item{lat, lon}{latitude and longitude as separate vectors, or as
data.frame with corresponding column names. The data.frame is
passed as first argument to \code{toMODISxy}, i.e. formally as
parameter \code{lat}.}

\item{grid}{MODIS
grid size (250, 500, or 1000; defaults to 250).}

\item{gx, gy}{MODIS global coordinates as separate vectors, or as
data.frame with corresponding column names. The data frame is
passed as first argument to \code{fromMODISxy}, i.e. formally
as argument \code{gx}. The global coordinates are line/sample
data but continuous across tiles.}
}
\value{
Returns a data.frame with MODIS coordinates (including tile
    number and sample/line within tile) or latitude/longitude.
}
\description{
These functions convert MODIS line/sample coordinates (250, 500 or
1000m pixel size) to latitude/longitude coordinates, assuming a
sinusoidal projection.  These calculations are equivalent to the
ones offered on-line by the MODLAND tile calculator (see link below).
}
\examples{
require(pgeo)
## let's take a test location near Bern, Switzerland
lat <- 46.95104
lon <- 7.439135
r <- toMODISxy(lat,lon)
r
##      gx    gy tileh tilev sample line
## 1 88837 20663    18     4   2437 1463

fromMODISxy(r)
##        lat      lon
## 1 46.95104 7.439135

## Compare this to the tile mapper output for the forward mapping of this coordinate:
## http://landweb.nascom.nasa.gov/cgi-bin/developer/tilemap.cgi
## Map Projection/Grid: Sinusoidal
## Pixel size: 0.25km
## Mapping type: Tile/image coordinates 
## Forward mapping (geographic coordinates in degrees): latitude, longitude
## [sn q fwd tp] lat 46.951040  long 7.439135  =>  vert tile 4  horiz tile 18  line 1463.00  samp 2437.00
}
\seealso{
\url{https://landweb.modaps.eosdis.nasa.gov/cgi-bin/developer/tilemap.cgi}
}
