% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethdf.R
\name{get_hdf_dirs}
\alias{get_hdf_dirs}
\alias{get_hdf_list}
\alias{get_hdf_file}
\title{Retrieve MODIS HDF file}
\usage{
get_hdf_dirs(
  sat = c("aqua", "terra"),
  years = c("0000", "9999"),
  version = 6,
  product = "13Q1",
  server = "https://ladsweb.modaps.eosdis.nasa.gov/archive/allData/",
  verbose = 1,
  cacheage = 24
)

get_hdf_list(
  tiles = c("h18v04"),
  sat = c("aqua", "terra"),
  years = c("0000", "9999"),
  version = 6,
  product = "13Q1",
  server = "https://ladsweb.modaps.eosdis.nasa.gov/archive/allData/",
  verbose = 1,
  cacheage = 24
)

get_hdf_file(
  url = NULL,
  file = NULL,
  outdir = NULL,
  appkey = NULL,
  verbose = 0,
  overwrite = FALSE
)
}
\arguments{
\item{sat}{the MODIS instruments for which data is retrieved,
e.g. \code{aqua} or \code{terra}, or both (default).}

\item{years}{an interval to which hdf file retrieval should be
restricted. Start and end years are provided as
\code{yyyy}.  To retrieve the entire data of year 2010,
one could use \code{years='2010'}.}

\item{version}{MODIS file version, currently 6 (default).}

\item{product}{product to retrieve, defaults to \code{13Q1}.}

\item{server}{server to use, including base path. defaults to
\code{https://ladsweb.modaps.eosdis.nasa.gov/archive/allData/}.}

\item{verbose}{integer indicating verbosity. Values supported are
0, 1 and 2}

\item{cacheage}{indicates whether a cache for the URL directories
should be used. Positive values indicate the maximum age in
hours.}

\item{tiles}{the MODIS tiles to download in format \code{h??v??}
(e.g. \code{h18v04})}

\item{url}{full url of MODIS file to be downloaded.}

\item{file}{if specifies, \code{file} allows to change the file
name under which the MODIS hdf file is saved.}

\item{outdir}{directory where the hdf file is to be stored. This
can be a relative or absolute path.}

\item{appkey}{App Key for data access, exported rom the LAADS DAAC
site (https://ladsweb.modaps.eosdis.nasa.gov/profile/#app-keys)}

\item{overwrite}{\code{get_hdf_file} skips the download of files
that already exist on disk, unless \code{overwrite=TRUE}.}
}
\value{
A data frame with columns \code{url}, \code{dir}, and
    \code{file}.
}
\description{
These functions retrieve MODIS Aqua and Terra hdf files and save
these locally.
}
\details{
\code{get_hdf_list} retrieves the URL of all MODIS hdf files from
Aqua, Terra, or both and returns them as a \code{data.frame}.  It
is possibly to specify a MODIS file version (currently 6, the
default) and to narrow down the list by specifying a year interval
(argument \code{years}).

\code{get_hdf_file} retrieves on or more HDF files.
\code{username} and \code{password} to access the server need to be
passed as arguments.  Data retrieval occurs using \code{curl},
which is invoked using \code{system}.  Note that username and
password are temporarily stored in a local 'netrc' file, which is
removed after the hdf file has been downloaded.  This 'netrc' file
is readable by the user only (\code{chmod 0600}) and removed after
the hdf file has been downloaded.
}
\examples{
\dontrun{
require(pgeo);
h <- get_hdf_list(yearss=c("2010","2015"))
get_hdf_files(h$url, addkey='xxxx')
}

}
